requirements.txt
httpx==0.13.3
protobuf==3.20.3
absl-py==2.2.2
adagio==0.2.6
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.11.1
anyio==4.6.2.post1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
bertopic==0.17.0
bleach==6.2.0
blinker==1.9.0
bokeh==2.4.3
cachetools==5.5.2
category_encoders==2.7.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
dacite==1.9.2
dash==3.0.3
dask==2024.2.1
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==0.8.3
fastjsonschema==2.20.0
filelock==3.18.0
findspark==2.0.1
flatbuffers==25.2.10
fonttools==4.55.0
fqdn==1.5.1
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.12.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==4.0.0rc1
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.71.0
h11==0.9.0
h2==3.2.0
h5py==3.13.0
hdbscan==0.8.40
hpack==3.0.0
hstspreload==2025.1.1
htmlmin==0.1.12
httpcore==0.9.1
huggingface-hub==0.33.1
HyperclassifierSearch==1.0
hyperframe==5.2.0
idna==2.10
ImageHash==4.3.1
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.29.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.0.3
joblib==1.3.2
json5==0.9.28
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaleido==0.2.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
kneed==0.8.5
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.44.0
locket==1.0.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
Metaphone==0.6
missingno==0.5.2
mistune==3.0.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
multimethod==1.12
namex==0.0.9
narwhals==1.44.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
oauthlib==3.3.1
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.16
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patsy==1.0.1
phik==0.12.4
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
plotly-resampler==0.10.0
ply==3.11
pmdarima==2.0.4
prettytable==3.12.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.3.0
psutil==6.1.0
pure_eval==0.2.3
puremagic==1.28
py4j==0.10.9.9
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pydot==1.4.2
Pygments==2.18.0
pynndescent==0.5.13
pyod==2.0.4
pyparsing==3.2.0
PySocks==1.7.1
pyspark==4.0.0
python-crfsuite==0.9.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-stdnum==1.20
pytz==2024.2
pyvi==0.1.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.2.0
qpd==0.4.4
rapidfuzz==2.15.2
referencing==0.35.1
regex==2021.11.10
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.21.0
rsa==4.9.1
safetensors==0.5.3
schemdraw==0.15
scikit-base==0.7.8
scikit-learn>=1.0.0,<1.6.0
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
selenium==4.27.1
Send2Trash==1.8.3
sentence-transformers==4.1.0
six==1.16.0
sklearn-compat==0.1.3
sklearn-crfsuite==0.5.0
sktime==0.26.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.5.1
statsmodels==0.14.4
streamlit==1.46.0
sympy==1.14.0
tabulate==0.9.0
tangled-up-in-unicode==0.1.0
tbats==1.1.3
tenacity==9.0.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.1.0
toolz==1.0.0
torch==2.7.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triad==0.9.8
trio==0.28.0
trio-websocket==0.11.1
tsdownsample==0.1.4.1
typeguard==4.4.2
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
underthesea==6.8.4
underthesea_core==1.0.4
uri-template==1.3.0
urllib3==2.2.3
varname==0.8.3
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
xgboost==3.0.0
xxhash==3.5.0
yarl==1.18.3
yellowbrick==1.5
zipp==3.21.0
